Revert "feat: FileInput component (internal) (#2912)"
This reverts commit 1a8e5d97912c66cb359fcf872acd6f44e26e0237.
