+++ src/file-upload/internal.tsx
@@ -12,7 +12,6 @@ import { useFormFieldContext } from '../contexts/form-field';
 import { ConstraintText, FormFieldError, FormFieldWarning } from '../form-field/internal';
 import { getBaseProps } from '../internal/base-component';
 import InternalFileDropzone, { useFilesDragging } from '../internal/components/file-dropzone';
-import InternalFileInput from '../internal/components/file-input';
 import TokenList from '../internal/components/token-list';
 import { fireNonCancelableEvent } from '../internal/events';
 import checkControlled from '../internal/hooks/check-controlled';
@@ -23,11 +22,12 @@ import { useUniqueId } from '../internal/hooks/use-unique-id';
 import { joinStrings } from '../internal/utils/strings';
 import InternalSpaceBetween from '../space-between/internal';
 import { Token } from '../token-group/token';
+import FileInput from './file-input';
 import { FileOption } from './file-option';
 import { FileUploadProps } from './interfaces';
 
-import fileInputStyles from '../internal/components/file-input/styles.css.js';
 import tokenListStyles from '../internal/components/token-list/styles.css.js';
+import fileInputStyles from './file-input/styles.css.js';
 import styles from './styles.css.js';
 
 type InternalFileUploadProps = FileUploadProps & InternalBaseComponentProps;
@@ -65,7 +65,7 @@ function InternalFileUpload(
     },
     listItemSelector: `.${tokenListStyles['list-item']}`,
     showMoreSelector: `.${tokenListStyles.toggle}`,
-    fallbackSelector: `.${fileInputStyles['file-input']}`,
+    fallbackSelector: `.${fileInputStyles['upload-input']}`,
   });
 
   const baseProps = getBaseProps(restProps);
@@ -128,19 +128,19 @@ function InternalFileUpload(
             {i18nStrings.dropzoneText(multiple)}
           </InternalFileDropzone>
         ) : (
-          <InternalFileInput
+          <FileInput
             ref={ref}
             accept={accept}
             ariaRequired={ariaRequired}
             multiple={multiple}
-            onChange={event => handleFilesChange(event.detail.value)}
+            onChange={handleFilesChange}
             value={value}
             {...restProps}
             ariaDescribedby={ariaDescribedBy}
             invalid={invalid}
           >
             {i18nStrings.uploadButtonText(multiple)}
-          </InternalFileInput>
+          </FileInput>
         )}
 
         {(constraintText || errorText || warningText) && (
